+++ src/api/paged/lazy.rs
@@ -4,33 +4,173 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use std::sync::{Arc, Mutex};
+use std::sync::RwLock;
 
-use async_trait::async_trait;
 use http::{header, Request};
 use serde::de::DeserializeOwned;
+use url::Url;
 
 use crate::api::paged::link_header;
-use crate::api::{
-    query, ApiError, AsyncClient, AsyncQuery, Client, Endpoint, Pageable, Pagination, Query,
-};
+use crate::api::{query, ApiError, Client, Endpoint, Pageable, Paged, Query, RestClient};
 
-/// A query modifier that paginates an endpoint.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct Paged<E> {
-    pub(in crate::api::paged) endpoint: E,
-    pub(in crate::api::paged) pagination: Pagination,
-}
-
-/// Collect data from a paged endpoint.
-pub fn paged<E>(endpoint: E, pagination: Pagination) -> Paged<E> {
-    Paged {
-        endpoint,
-        pagination,
+impl<E> Paged<E>
+where
+    E: Endpoint,
+    E: Pageable,
+{
+    /// Create an iterator over the results of paginated results for with a client.
+    pub fn iter<'a, C, T>(&'a self, client: &'a C) -> LazilyPagedIter<'a, E, C, T> {
+        LazilyPagedIter::new(self, client)
     }
 }
 
-impl<E, T, C> Query<Vec<T>, C> for Paged<E>
+#[derive(Debug, Clone, PartialEq, Eq)]
+enum KeysetPage {
+    First,
+    Next(Url),
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+enum Page {
+    Number(u64),
+    Keyset(KeysetPage),
+    Done,
+}
+
+impl Page {
+    fn next_url(&self) -> Option<&Url> {
+        if let Self::Keyset(KeysetPage::Next(url)) = self {
+            Some(url)
+        } else {
+            None
+        }
+    }
+
+    fn next_page(&mut self, next_url: Option<Url>) {
+        let next_page = match *self {
+            Self::Number(page) => Self::Number(page + 1),
+            Self::Keyset(_) => {
+                if let Some(next_url) = next_url {
+                    Self::Keyset(KeysetPage::Next(next_url))
+                } else {
+                    Self::Done
+                }
+            },
+            Self::Done => Self::Done,
+        };
+
+        *self = next_page;
+    }
+
+    fn apply_to(&self, pairs: &mut url::form_urlencoded::Serializer<url::UrlQuery>) {
+        match self {
+            Self::Number(page) => {
+                let page_str = format!("{}", page);
+                pairs.append_pair("page", &page_str);
+            },
+            Self::Keyset(_) => {
+                pairs.append_pair("pagination", "keyset");
+            },
+            Self::Done => {
+                unreachable!("The `Done` state should not be applied to any url")
+            },
+        }
+    }
+}
+
+struct PageState {
+    total_results: usize,
+    next_page: Page,
+}
+
+struct LazilyPagedState<'a, E> {
+    paged: &'a Paged<E>,
+    page_state: RwLock<PageState>,
+}
+
+impl<'a, E> LazilyPagedState<'a, E>
+where
+    E: Pageable,
+{
+    fn new(paged: &'a Paged<E>) -> Self {
+        let next_page = if paged.endpoint.use_keyset_pagination() {
+            Page::Keyset(KeysetPage::First)
+        } else {
+            Page::Number(1)
+        };
+
+        let page_state = PageState {
+            total_results: 0,
+            next_page,
+        };
+
+        Self {
+            paged,
+            page_state: RwLock::new(page_state),
+        }
+    }
+}
+
+impl<'a, E> LazilyPagedState<'a, E> {
+    fn next_page(&self, last_page_size: usize, next_url: Option<Url>) {
+        let mut page_state = self.page_state.write().expect("poisoned next_page");
+        page_state.total_results += last_page_size;
+
+        // Gitlab used to have issues returning paginated results; these have been fixed since, but
+        // if it is needed, the bug manifests as Gitlab returning *all* results instead of just the
+        // requested results. This can cause an infinite loop here if the number of total results
+        // is exactly equal to `per_page`.
+        if self
+            .paged
+            .pagination
+            .is_last_page(last_page_size, page_state.total_results)
+        {
+            page_state.next_page = Page::Done;
+        } else {
+            page_state.next_page.next_page(next_url);
+        }
+    }
+}
+
+impl<'a, E> LazilyPagedState<'a, E>
+where
+    E: Endpoint,
+{
+    fn page_url<C>(&self, client: &C) -> Result<Option<Url>, ApiError<C::Error>>
+    where
+        C: RestClient,
+    {
+        let page_state = self.page_state.read().expect("poisoned next_page");
+        let next_page = &page_state.next_page;
+
+        if *next_page == Page::Done {
+            return Ok(None);
+        }
+
+        let url = if let Some(next_url) = next_page.next_url() {
+            next_url.clone()
+        } else {
+            let mut url = client.rest_endpoint(&self.paged.endpoint.endpoint())?;
+            self.paged.endpoint.parameters().add_to_url(&mut url);
+
+            let per_page = self.paged.pagination.page_limit();
+            let per_page_str = format!("{}", per_page);
+
+            {
+                let mut pairs = url.query_pairs_mut();
+                pairs.append_pair("per_page", &per_page_str);
+
+                next_page.apply_to(&mut pairs);
+            }
+
+            url
+        };
+
+        Ok(Some(url))
+    }
+}
+
+impl<'a, E, T, C> Query<Vec<T>, C> for LazilyPagedState<'a, E>
 where
     E: Endpoint,
     E: Pageable,
@@ -38,105 +178,93 @@ where
     C: Client,
 {
     fn query(&self, client: &C) -> Result<Vec<T>, ApiError<C::Error>> {
-        self.iter(client).collect()
+        let url = if let Some(url) = self.page_url(client)? {
+            url
+        } else {
+            // Just return empty data.
+            // XXX: Return a new kind of PaginationError here?
+            return Ok(Vec::new());
+        };
+
+        let body = self.paged.endpoint.body()?;
+
+        let req = Request::builder()
+            .method(self.paged.endpoint.method())
+            .uri(query::url_to_http_uri(url));
+        let (req, data) = if let Some((mime, data)) = body.as_ref() {
+            let req = req.header(header::CONTENT_TYPE, *mime);
+            (req, data.clone())
+        } else {
+            (req, Vec::new())
+        };
+        let rsp = client.rest(req, data)?;
+        let status = rsp.status();
+
+        let next_url = if self.paged.endpoint.use_keyset_pagination() {
+            link_header::next_page_from_headers(rsp.headers())?
+        } else {
+            None
+        };
+
+        let v = serde_json::from_slice(rsp.body())?;
+        if !status.is_success() {
+            return Err(ApiError::from_gitlab(v));
+        }
+
+        let page = serde_json::from_value::<Vec<T>>(v).map_err(ApiError::data_type::<Vec<T>>)?;
+        self.next_page(page.len(), next_url);
+
+        Ok(page)
     }
 }
 
-#[async_trait]
-impl<E, T, C> AsyncQuery<Vec<T>, C> for Paged<E>
+/// An iterator which yields items from a paginated result.
+///
+/// The pages are fetched lazily, so endpoints not using keyset pagination may observe duplicate or
+/// missing items (depending on sorting) if new objects are created or removed while iterating.
+pub struct LazilyPagedIter<'a, E, C, T> {
+    client: &'a C,
+    state: LazilyPagedState<'a, E>,
+    current_page: Vec<T>,
+}
+
+impl<'a, E, C, T> LazilyPagedIter<'a, E, C, T>
 where
-    E: Endpoint + Sync,
+    E: Endpoint,
     E: Pageable,
-    T: DeserializeOwned + Send + 'static,
-    C: AsyncClient + Sync,
 {
-    async fn query_async(&self, client: &C) -> Result<Vec<T>, ApiError<C::Error>> {
-        let url = {
-            let mut url = client.rest_endpoint(&self.endpoint.endpoint())?;
-            self.endpoint.parameters().add_to_url(&mut url);
-            url
-        };
+    fn new(paged: &'a Paged<E>, client: &'a C) -> Self {
+        let state = LazilyPagedState::new(paged);
 
-        let mut page_num = 1;
-        let per_page = self.pagination.page_limit();
-        let per_page_str = format!("{}", per_page);
+        Self {
+            client,
+            state,
+            current_page: Vec::new(),
+        }
+    }
+}
 
-        let results = Arc::new(Mutex::new(Vec::new()));
-        let mut next_url = None;
-        let use_keyset_pagination = self.endpoint.use_keyset_pagination();
+impl<'a, E, C, T> Iterator for LazilyPagedIter<'a, E, C, T>
+where
+    E: Endpoint,
+    E: Pageable,
+    T: DeserializeOwned,
+    C: Client,
+{
+    type Item = Result<T, ApiError<C::Error>>;
 
-        let body = self.endpoint.body()?;
-
-        loop {
-            let page_url = if let Some(url) = next_url.take() {
-                url
-            } else {
-                let page_str = format!("{}", page_num);
-                let mut page_url = url.clone();
-
-                {
-                    let mut pairs = page_url.query_pairs_mut();
-                    pairs.append_pair("per_page", &per_page_str);
-
-                    if use_keyset_pagination {
-                        pairs.append_pair("pagination", "keyset");
-                    } else {
-                        pairs.append_pair("page", &page_str);
-                    }
-                }
-
-                page_url
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.current_page.is_empty() {
+            self.current_page = match self.state.query(self.client) {
+                Ok(data) => data,
+                Err(err) => return Some(Err(err)),
             };
 
-            let req = Request::builder()
-                .method(self.endpoint.method())
-                .uri(query::url_to_http_uri(page_url));
-            let (req, data) = if let Some((mime, data)) = body.as_ref() {
-                let req = req.header(header::CONTENT_TYPE, *mime);
-                (req, data.clone())
-            } else {
-                (req, Vec::new())
-            };
-            let rsp = client.rest_async(req, data).await?;
-            let status = rsp.status();
-
-            if use_keyset_pagination {
-                next_url = link_header::next_page_from_headers(rsp.headers())?;
-            }
-
-            let v = serde_json::from_slice(rsp.body())?;
-            if !status.is_success() {
-                return Err(ApiError::from_gitlab(v));
-            }
-
-            let page =
-                serde_json::from_value::<Vec<T>>(v).map_err(ApiError::data_type::<Vec<T>>)?;
-            let page_len = page.len();
-
-            // Gitlab used to have issues returning paginated results; these have been fixed since,
-            // but if it is needed, the bug manifests as Gitlab returning *all* results instead of
-            // just the requested results. This can cause an infinite loop here if the number of
-            // total results is exactly equal to `per_page`.
-            let is_last_page = {
-                let mut locked_results = results.lock().expect("poisoned results");
-                locked_results.extend(page);
-                self.pagination.is_last_page(page_len, locked_results.len())
-            };
-            if is_last_page {
-                break;
-            }
-
-            if use_keyset_pagination {
-                if next_url.is_none() {
-                    break;
-                }
-            } else {
-                page_num += 1;
-            }
+            // Reverse the page order so that `.pop()` works.
+            self.current_page.reverse();
         }
 
-        let mut locked_results = results.lock().expect("poisoned results");
-        Ok(std::mem::take(&mut locked_results))
+        self.current_page.pop().map(Ok)
     }
 }
 
@@ -147,7 +275,7 @@ mod tests {
     use serde_json::json;
 
     use crate::api::endpoint_prelude::*;
-    use crate::api::{self, ApiError, AsyncQuery, Pagination, Query};
+    use crate::api::{self, ApiError, Pagination};
     use crate::test::client::{ExpectedUrl, PagedTestClient, SingleTestClient};
 
     #[derive(Debug, Default)]
@@ -186,7 +314,9 @@ mod tests {
         let client = SingleTestClient::new_raw(endpoint, "not json");
         let endpoint = Dummy::default();
 
-        let res: Result<Vec<DummyResult>, _> = api::paged(endpoint, Pagination::All).query(&client);
+        let res: Result<Vec<DummyResult>, _> = api::paged(endpoint, Pagination::All)
+            .iter(&client)
+            .collect();
         let err = res.unwrap_err();
         if let ApiError::Json {
             source,
@@ -209,7 +339,9 @@ mod tests {
         let client = SingleTestClient::new_raw(endpoint, "");
         let endpoint = Dummy::default();
 
-        let res: Result<Vec<DummyResult>, _> = api::paged(endpoint, Pagination::All).query(&client);
+        let res: Result<Vec<DummyResult>, _> = api::paged(endpoint, Pagination::All)
+            .iter(&client)
+            .collect();
         let err = res.unwrap_err();
         if let ApiError::Json {
             source,
@@ -240,7 +372,9 @@ mod tests {
         );
         let endpoint = Dummy::default();
 
-        let res: Result<Vec<DummyResult>, _> = api::paged(endpoint, Pagination::All).query(&client);
+        let res: Result<Vec<DummyResult>, _> = api::paged(endpoint, Pagination::All)
+            .iter(&client)
+            .collect();
         let err = res.unwrap_err();
         if let ApiError::Gitlab {
             msg,
@@ -268,7 +402,9 @@ mod tests {
         );
         let endpoint = Dummy::default();
 
-        let res: Result<Vec<DummyResult>, _> = api::paged(endpoint, Pagination::All).query(&client);
+        let res: Result<Vec<DummyResult>, _> = api::paged(endpoint, Pagination::All)
+            .iter(&client)
+            .collect();
         let err = res.unwrap_err();
         if let ApiError::Gitlab {
             msg,
@@ -294,7 +430,9 @@ mod tests {
         let client = SingleTestClient::new_json(endpoint, &err_obj);
         let endpoint = Dummy::default();
 
-        let res: Result<Vec<DummyResult>, _> = api::paged(endpoint, Pagination::All).query(&client);
+        let res: Result<Vec<DummyResult>, _> = api::paged(endpoint, Pagination::All)
+            .iter(&client)
+            .collect();
         let err = res.unwrap_err();
         if let ApiError::GitlabUnrecognized {
             obj,
@@ -326,36 +464,8 @@ mod tests {
         };
 
         let res: Vec<DummyResult> = api::paged(query, Pagination::Limit(25))
-            .query(&client)
-            .unwrap();
-        assert_eq!(res.len(), 25);
-        for (i, value) in res.iter().enumerate() {
-            assert_eq!(value.value, i as u8);
-        }
-    }
-
-    #[tokio::test]
-    async fn test_pagination_limit_async() {
-        let endpoint = ExpectedUrl::builder()
-            .endpoint("paged_dummy")
-            .paginated(true)
-            .build()
-            .unwrap();
-        let client = PagedTestClient::new_raw(
-            endpoint,
-            (0..=255).map(|value| {
-                DummyResult {
-                    value,
-                }
-            }),
-        );
-        let query = Dummy {
-            with_keyset: false,
-        };
-
-        let res: Vec<DummyResult> = api::paged(query, Pagination::Limit(25))
-            .query_async(&client)
-            .await
+            .iter(&client)
+            .collect::<Result<Vec<_>, _>>()
             .unwrap();
         assert_eq!(res.len(), 25);
         for (i, value) in res.iter().enumerate() {
@@ -380,33 +490,9 @@ mod tests {
         );
         let query = Dummy::default();
 
-        let res: Vec<DummyResult> = api::paged(query, Pagination::All).query(&client).unwrap();
-        assert_eq!(res.len(), 256);
-        for (i, value) in res.iter().enumerate() {
-            assert_eq!(value.value, i as u8);
-        }
-    }
-
-    #[tokio::test]
-    async fn test_pagination_all_async() {
-        let endpoint = ExpectedUrl::builder()
-            .endpoint("paged_dummy")
-            .paginated(true)
-            .build()
-            .unwrap();
-        let client = PagedTestClient::new_raw(
-            endpoint,
-            (0..=255).map(|value| {
-                DummyResult {
-                    value,
-                }
-            }),
-        );
-        let query = Dummy::default();
-
         let res: Vec<DummyResult> = api::paged(query, Pagination::All)
-            .query_async(&client)
-            .await
+            .iter(&client)
+            .collect::<Result<Vec<_>, _>>()
             .unwrap();
         assert_eq!(res.len(), 256);
         for (i, value) in res.iter().enumerate() {
@@ -434,7 +520,8 @@ mod tests {
         };
 
         let res: Vec<DummyResult> = api::paged(query, Pagination::Limit(25))
-            .query(&client)
+            .iter(&client)
+            .collect::<Result<Vec<_>, _>>()
             .unwrap();
         assert_eq!(res.len(), 25);
         for (i, value) in res.iter().enumerate() {
@@ -461,7 +548,10 @@ mod tests {
             with_keyset: true,
         };
 
-        let res: Vec<DummyResult> = api::paged(query, Pagination::All).query(&client).unwrap();
+        let res: Vec<DummyResult> = api::paged(query, Pagination::All)
+            .iter(&client)
+            .collect::<Result<Vec<_>, _>>()
+            .unwrap();
         assert_eq!(res.len(), 256);
         for (i, value) in res.iter().enumerate() {
             assert_eq!(value.value, i as u8);
@@ -488,7 +578,8 @@ mod tests {
         };
 
         let res: Vec<DummyResult> = api::paged(query, Pagination::Limit(300))
-            .query(&client)
+            .iter(&client)
+            .collect::<Result<Vec<_>, _>>()
             .unwrap();
         assert_eq!(res.len(), 256);
         for (i, value) in res.iter().enumerate() {